Combined journal_entry.dart and mod_entry.dart
